implementata la votabilità degli interventi (con interfaccia ajax)
git-svn-id: http://svn.openpolis.it/repos/openparlamento/trunk@1416 f338ed83-fc16-40d3-870d-fec3a6a98b5b

+++ plugins/deppPropelActAsVotableBehaviorPlugin/lib/deppPropelActAsVotableBehavior.class.php
@@ -216,11 +216,11 @@ class deppPropelActAsVotableBehavior
     $c->add(sfVotingPeer::VOTABLE_MODEL, get_class($object));
     $c->add(sfVotingPeer::USER_ID, $user_id);
     $v = sfVotingPeer::doSelectOne($c);
-    $v->delete();
+    if ($v instanceof sfVoting) {
+      $v->delete();
+    }
     self::setVotingToObject($object, $this->getVoting($object, self::getPrecision(), true));
     self::setVotingDetailsToObject($object, $this->getVotingDetails($object, true, true));
-
-    return $ret;
   }
 
   /**
